Finding and drawing critical light rays
To create light rays, I found the vertices for rectangular ground objects, and drew lines from the player to each vertex on each object.

Some of those lines will be the light rays between light and dark.

Next step will be to find which lines are actually the light ray vs lines that are just in the area of light.
(easier to understand once its finished)