Found only relevant light rays
I further found relevant light rays by determining if the light ray would hit a solid surface.
To do this, I took the angle bisectors between the two edges and the light ray, and took the dot product between them to get the angle. If the angle is greater than 45 degrees, then it hits a solid wall and is not a relevant light ray.

Iterations:
- Tried checking if the dot products between the two edges and the light ray were both under 180 degrees, which would mean that it is solid. Didn't work because dot product doesn't give angles above 180.
- Struggled with lines flashing in and out, was because of floating point error comparing to a constant pi/4

Misc:
- Removed gravity for the player to make it easier to test
- Showed collision boxes and removed icon for better visibility
- Made some testing drawings to visualize what was going on